requirements.txt
TTS>=0.22.0
numpy>=1.24.0
scipy>=1.10.0
soundfile>=0.12.0
librosa>=0.10.0
pydub>=0.25.1
python-telegram-bot>=20.0
python-dotenv>=1.0.0
